clean a bit the code
