--- plugins/polio/js/src/domains/Calendar/campaignCalendar/utils.tsx
@@ -1,944 +1,0 @@
-import React, { ReactElement } from 'react';
-import moment, { Moment } from 'moment';
-
-import { CalendarCampaign } from '../../../constants/types';
-import { CampaignDurationCell } from './cells/CampaignDuration';
-import { EmptyCell } from './cells/Empty';
-import { RoundCell } from './cells/RoundCell';
-import { SubactivityCell } from './cells/SubactivityCell';
-import { colsCounts, dateFormat } from './constants';
-import {
-    CalendarData,
-    CalendarRound,
-    MappedCampaign,
-    MonthHeader,
-    PeriodType,
-    SubActivity,
-    WeekHeader,
-    YearHeader,
-} from './types';
-
-const getEmptyCellsData = (
-    endDate: Moment,
-    lastSunday: Moment,
-): { extraDays: number; fullWeeks: number } => {
-    let sunday: Moment;
-    if (endDate.isAfter(lastSunday)) {
-        return { extraDays: 0, fullWeeks: 0 };
-    }
-    if (endDate.weekday() !== 7) {
-        sunday = endDate.clone().endOf('isoWeek');
-    } else {
-        sunday = endDate.clone();
-    }
-    const extraDays = sunday.clone().diff(endDate, 'days');
-    const fullWeeks = lastSunday.diff(sunday, 'weeks');
-    return {
-        extraDays,
-        fullWeeks,
-    };
-};
-const getEmptyCells = ({
-    cells,
-    extraDays,
-    fullWeeks,
-    currentWeekIndex,
-    id,
-}: {
-    cells: ReactElement[];
-    extraDays: number;
-    fullWeeks: number;
-    currentWeekIndex: number;
-    id: string;
-}) => {
-    let spans = 0;
-    cells.forEach(c => {
-        spans += c.props.colSpan;
-    });
-    spans = Math.floor(spans / 7);
-    if (extraDays) {
-        spans += 1;
-        cells.push(
-            <EmptyCell
-                key={`empty-cell-${id}-end`}
-                colSpan={extraDays}
-                isCurrentWeek={spans === currentWeekIndex}
-            />,
-        );
-    }
-    const weeks = fullWeeks >= 0 ? fullWeeks : 0;
-    Array(weeks)
-        .fill(null)
-        .forEach((_, i) => {
-            spans += 1;
-            cells.push(
-                <EmptyCell
-                    key={`empty-cell-week-${id}-end-${i}-${spans}`}
-                    colSpan={7}
-                    isCurrentWeek={spans === currentWeekIndex}
-                />,
-            );
-        });
-    return cells;
-};
-
-const getCalendarData = (
-    currentMonday: Moment,
-    periodType: PeriodType = 'year',
-): CalendarData => {
-    const todayMonday = moment().startOf('isoWeek');
-    const currentWeek = {
-        year: todayMonday.format('YYYY'),
-        month: todayMonday.format('MMM'),
-        value: todayMonday.format('DD'),
-    };
-    const columnCount = colsCounts[periodType];
-    const firstMonday = currentMonday
-        .clone()
-        .subtract(Math.floor(columnCount / 2), 'week');
-    const lastSunday = currentMonday
-        .clone()
-        .add(columnCount - (Math.floor(columnCount / 2) + 1), 'week')
-        .endOf('isoWeek');
-    const headers: {
-        years: YearHeader[];
-        months: MonthHeader[];
-        weeks: WeekHeader[];
-    } = {
-        years: [],
-        months: [],
-        weeks: [],
-    };
-    let currentWeekIndex = -1;
-    Array(columnCount)
-        .fill(null)
-        .forEach((_, i) => {
-            const newMonday = firstMonday.clone().add(i, 'week');
-            const year = newMonday.format('YYYY');
-            if (!headers.years.find(y => y.value === year)) {
-                headers.years.push({
-                    value: year,
-                    daysCount: 0,
-                });
-            }
-            const month = newMonday.format('MMM');
-            if (
-                !headers.months.find(m => month === m.value && year === m.year)
-            ) {
-                headers.months.push({
-                    year,
-                    value: month,
-                    daysCount: 0,
-                });
-            }
-            const week = newMonday.format('DD');
-            if (
-                currentWeek.year === year &&
-                currentWeek.month === month &&
-                currentWeek.value === week
-            ) {
-                currentWeekIndex = i + 1;
-            }
-            const monthIndex = headers.months.findIndex(
-                m => month === m.value && year === m.year,
-            );
-            headers.months[monthIndex].daysCount += 7;
-            const yearIndex = headers.years.findIndex(y => year === y.value);
-            headers.years[yearIndex].daysCount += 7;
-            headers.weeks.push({
-                year,
-                month,
-                value: week,
-                monday: newMonday.clone(),
-            });
-        });
-    return {
-        headers,
-        currentWeekIndex,
-        firstMonday,
-        lastSunday,
-    };
-};
-
-const isDateInRange = (date: Moment, firstMonday: Moment, lastSunday: Moment) =>
-    Boolean(date) &&
-    date.isSameOrBefore(lastSunday, 'day') &&
-    date.isSameOrAfter(firstMonday, 'day');
-
-const getLastRound = (rounds: CalendarRound[]): CalendarRound => {
-    const lastRound = rounds[rounds.length - 1];
-    if (lastRound?.started_at && lastRound?.ended_at) {
-        return lastRound;
-    }
-    const lastRoundRemoved = [...rounds].filter((_rnd, index) => {
-        return index !== rounds.length - 1;
-    });
-    if (lastRoundRemoved.length >= 1) {
-        return getLastRound(lastRoundRemoved);
-    }
-    return lastRound;
-};
-const filterCampaigns = (
-    allCampaigns: MappedCampaign[],
-    firstMonday: Moment,
-    lastSunday: Moment,
-): MappedCampaign[] => {
-    return allCampaigns.filter(campaign => {
-        const { rounds, subActivities } = campaign;
-        const lastRound = getLastRound(rounds);
-        const endCampaignDate = lastRound?.end
-            ?.clone()
-            .add(lastRound.daysCount, 'days');
-
-        const isRoundVisible =
-            rounds.filter(
-                round =>
-                    round.started_at &&
-                    round.ended_at &&
-                    round.start &&
-                    round.end &&
-                    (isDateInRange(round.start, firstMonday, lastSunday) ||
-                        isDateInRange(round.end, firstMonday, lastSunday)),
-            ).length > 0;
-
-        const isLastRoundVisible = endCampaignDate
-            ? isDateInRange(endCampaignDate, firstMonday, lastSunday)
-            : false;
-
-        // const isSubActivityVisible = subActivities.some(
-        //     subActivity =>
-        //         (subActivity.start_date &&
-        //             isDateInRange(
-        //                 subActivity.start_date,
-        //                 firstMonday,
-        //                 lastSunday,
-        //             )) ||
-        //         (subActivity.end_date &&
-        //             isDateInRange(
-        //                 subActivity.end_date,
-        //                 firstMonday,
-        //                 lastSunday,
-        //             )),
-        // );
-
-        // return isRoundVisible || isLastRoundVisible || isSubActivityVisible;
-        return isRoundVisible || isLastRoundVisible;
-    });
-};
-
-const mapCampaigns = (
-    allCampaigns: CalendarCampaign[],
-    firstMonday: Moment,
-    lastSunday: Moment,
-): MappedCampaign[] => {
-    return allCampaigns.map(c => {
-        const rounds = c.rounds.map((round, index) => {
-            const nextRound = c.rounds[index + 1];
-            const started_at =
-                round.started_at === null ? undefined : round.started_at;
-            const ended_at =
-                round.ended_at === null ? undefined : round.ended_at;
-            const start = round.started_at
-                ? moment(round.started_at, dateFormat)
-                : undefined;
-            const end = round.ended_at
-                ? moment(round.ended_at, dateFormat)
-                : undefined;
-            const hasNextRound = nextRound && end && nextRound.started_at;
-            const target_population =
-                round.target_population != null ? round.target_population : 0;
-            const weeksCount = hasNextRound
-                ? moment(nextRound.started_at, dateFormat).diff(end, 'weeks')
-                : 0;
-
-            const daysCount = hasNextRound
-                ? moment(nextRound.started_at, dateFormat).diff(end, 'days')
-                : 0;
-
-            const hasSubActivities = Boolean(
-                c.sub_activities.find(
-                    subActivity => subActivity.round_number === round.number,
-                ),
-            );
-
-            const mappedRound = {
-                ...round,
-                target_population,
-                started_at,
-                ended_at,
-                start,
-                end,
-                weeksCount,
-                daysCount,
-                hasSubActivities,
-            };
-            if (c.is_preventive && !nextRound?.started_at) {
-                return { ...mappedRound, weeksCount: 0, daysCount: 0 };
-            }
-            return mappedRound;
-        });
-        const displayedSubActivities = c.sub_activities
-            .sort(
-                (a, b) =>
-                    new Date(a.start_date).getTime() -
-                    new Date(b.start_date).getTime(),
-            )
-            .map(subActivity => {
-                return {
-                    ...subActivity,
-                    start_date: moment(subActivity.start_date)
-                        .hours(0)
-                        .minutes(1),
-                    end_date: moment(subActivity.end_date)
-                        .hours(23)
-                        .minutes(59),
-                };
-            })
-            .filter(
-                subActivity =>
-                    isDateInRange(
-                        moment(subActivity.start_date),
-                        firstMonday,
-                        lastSunday,
-                    ) ||
-                    isDateInRange(
-                        moment(subActivity.end_date),
-                        firstMonday,
-                        lastSunday,
-                    ),
-            );
-
-        return {
-            id: c.id,
-            name: c.obr_name,
-            country: c.top_level_org_unit_name,
-            country_id: c.top_level_org_unit_id,
-            rounds,
-            original: c,
-            isPreventive: c.is_preventive,
-            isTest: c.is_test,
-            onHold: c.on_hold,
-            separateScopesPerRound: c.separate_scopes_per_round,
-            scopes: c.scopes,
-            subActivities: displayedSubActivities,
-            hasSubActivities: displayedSubActivities.length > 0,
-        };
-    });
-};
-
-const drawEmptyFirstCells = ({
-    startDate,
-    firstMonday,
-    currentWeekIndex,
-    id,
-}: {
-    startDate: Moment;
-    firstMonday: Moment;
-    currentWeekIndex: number;
-    id: string;
-}): ReactElement[] => {
-    const result: ReactElement[] = [];
-    if (!startDate.isSame(firstMonday, 'day')) {
-        const monday = startDate.clone().startOf('isoWeek');
-        const extraDays = startDate.clone().diff(monday, 'days');
-        const fullWeeks = monday.diff(firstMonday, 'weeks');
-        Array(fullWeeks)
-            .fill(null)
-            .forEach((_, i) => {
-                result.push(
-                    <EmptyCell
-                        key={`empty-cell-${id}-start-${i}`}
-                        colSpan={7}
-                        isCurrentWeek={result.length + 1 === currentWeekIndex}
-                    />,
-                );
-            });
-        if (extraDays) {
-            result.push(
-                <EmptyCell
-                    key={`empty-cell-${id}-start-${startDate}`}
-                    colSpan={extraDays}
-                    isCurrentWeek={result.length + 1 === currentWeekIndex}
-                />,
-            );
-        }
-    }
-    return result;
-};
-
-const addCellsBeforeRound = ({
-    id,
-    cells,
-    currentWeekIndex,
-    firstMonday,
-    lastSunday,
-    round,
-    index,
-    previousRound,
-    startInRange,
-    periodType,
-}: {
-    id: string;
-    cells: ReactElement[];
-    currentWeekIndex: number;
-    firstMonday: Moment;
-    lastSunday: Moment;
-    round: CalendarRound;
-    index: number;
-    previousRound?: CalendarRound;
-    startInRange: boolean;
-    periodType: PeriodType;
-}) => {
-    const result = [...cells];
-    if (index === 0) {
-        // if first round is not starting on the first cell, populate with empty cell
-        if (startInRange && round.start) {
-            const emptyCells = drawEmptyFirstCells({
-                startDate: round.start,
-                firstMonday,
-                currentWeekIndex,
-                id,
-            });
-            result.push(...emptyCells);
-        }
-    } else if (
-        // if previous round end date is not in range, calculate number of cells to firstMonday and draw campaign duration cells
-        previousRound?.end &&
-        !isDateInRange(previousRound.end, firstMonday, lastSunday) &&
-        startInRange &&
-        round.start
-    ) {
-        const campaignDays = round.start.diff(firstMonday, 'days');
-        result.push(
-            <CampaignDurationCell
-                key={`campaign-duration-${id}-round-${round.id}`}
-                colSpan={campaignDays}
-                weeksCount={previousRound.weeksCount || 0}
-                periodType={periodType}
-            />,
-        );
-    }
-    return result;
-};
-
-const addRoundCell = ({
-    campaign,
-    cells,
-    round,
-    firstMonday,
-    lastSunday,
-    startInRange,
-    endInRange,
-    periodType,
-}: {
-    campaign: MappedCampaign;
-    cells: ReactElement[];
-    round: CalendarRound;
-    firstMonday: Moment;
-    lastSunday: Moment;
-    startInRange: boolean;
-    endInRange: boolean;
-    periodType: PeriodType;
-}) => {
-    let colSpan = 1;
-    const { id } = campaign;
-    const result = [...cells];
-
-    const startAndEndInRange = startInRange && endInRange;
-
-    const onlyStartInRange = !endInRange && startInRange;
-
-    const onlyEndInRange = !startInRange && endInRange;
-
-    if (round.end && round.start) {
-        if (startAndEndInRange) {
-            // if both start and end date are in range use diff between dates for length of cells
-            colSpan = round.end.clone().add(1, 'day').diff(round.start, 'days');
-            // else if start is not in range calculate diff with firstmonday
-        } else if (onlyEndInRange) {
-            colSpan = round.end.clone().add(1, 'day').diff(firstMonday, 'days');
-            // else if end is not in range calculate diff with lastSunday
-        } else if (onlyStartInRange) {
-            colSpan = Math.abs(
-                round.start.clone().subtract(1, 'day').diff(lastSunday, 'days'),
-            );
-        }
-    }
-    result.push(
-        <RoundCell
-            key={`round${round.number}Cost-campaign-${id}`}
-            colSpan={colSpan}
-            campaign={campaign}
-            round={round}
-            periodType={periodType}
-        />,
-    );
-    return result;
-};
-const addSubactivityCell = ({
-    cells,
-    subActivity,
-    firstMonday,
-    lastSunday,
-    startInRange,
-    endInRange,
-}: {
-    cells: ReactElement[];
-    subActivity: SubActivity;
-    firstMonday: Moment;
-    lastSunday: Moment;
-    startInRange: boolean;
-    endInRange: boolean;
-}) => {
-    let colSpan = 1;
-    const result = [...cells];
-
-    const startAndEndInRange = startInRange && endInRange;
-
-    const onlyStartInRange = !endInRange && startInRange;
-
-    const onlyEndInRange = !startInRange && endInRange;
-
-    if (subActivity.start_date && subActivity.end_date) {
-        if (startAndEndInRange) {
-            // if both start and end date are in range use diff between dates for length of cells
-            colSpan = subActivity.end_date
-                .clone()
-                .add(1, 'day')
-                .diff(subActivity.start_date, 'days');
-            // else if start is not in range calculate diff with firstmonday
-        } else if (onlyEndInRange) {
-            colSpan = subActivity.end_date
-                .clone()
-                .add(1, 'day')
-                .diff(firstMonday, 'days');
-            // else if end is not in range calculate diff with lastSunday
-        } else if (onlyStartInRange) {
-            colSpan = Math.abs(
-                subActivity.start_date
-                    .clone()
-                    .subtract(1, 'day')
-                    .diff(lastSunday, 'days'),
-            );
-        }
-    }
-    result.push(
-        <SubactivityCell
-            key={`subactivity-${subActivity.id}`}
-            colSpan={colSpan}
-        />,
-    );
-    return result;
-};
-const addBufferCell = ({
-    id,
-    round,
-    cells,
-    dateUntilNextRoundInRange,
-    lastSunday,
-    endInRange,
-    dateUntilNextRound,
-    firstMonday,
-    rounds,
-    index,
-    periodType,
-}: {
-    id: string;
-    round: CalendarRound;
-    cells: ReactElement[];
-    dateUntilNextRoundInRange: boolean;
-    lastSunday: Moment;
-    endInRange: boolean;
-    dateUntilNextRound?: Moment;
-    firstMonday: Moment;
-    rounds: CalendarRound[];
-    index: number;
-    periodType: PeriodType;
-}) => {
-    // draw campaign duration cells
-    const result = [...cells];
-    let roundInterval =
-        round.daysCount && round.daysCount > 0 ? round.daysCount - 1 : 0;
-    // if campaign days count is not in range calculate diff with last sunday campaign
-    if (!dateUntilNextRoundInRange) {
-        roundInterval = lastSunday.clone().diff(round.end, 'days');
-    }
-    // if last round end date is not in range calculate diff with first monday and campaign days
-    if (!endInRange && dateUntilNextRoundInRange && dateUntilNextRound) {
-        roundInterval = dateUntilNextRound
-            .clone()
-            .add(1, 'day')
-            .diff(firstMonday, 'days');
-    }
-    const nextRound = rounds[index + 1];
-    const hasNextRound = Boolean(
-        nextRound && nextRound.started_at && nextRound.ended_at,
-    );
-    if (hasNextRound) {
-        result.push(
-            <CampaignDurationCell
-                key={`campaign-duration-${id}-round-${round.number}`}
-                colSpan={roundInterval}
-                weeksCount={round.weeksCount || 0}
-                periodType={periodType}
-            />,
-        );
-    }
-    return result;
-};
-
-const addRemainingEmptyCells = ({
-    cells,
-    dateUntilNextRound,
-    lastSunday,
-    campaign,
-    currentWeekIndex,
-}: {
-    cells: ReactElement[];
-    dateUntilNextRound: Moment;
-    lastSunday: Moment;
-    campaign: MappedCampaign;
-    currentWeekIndex: number;
-}) => {
-    const { id } = campaign;
-    const cellsData = getEmptyCellsData(dateUntilNextRound, lastSunday);
-    const { extraDays, fullWeeks } = cellsData;
-
-    return getEmptyCells({
-        cells,
-        extraDays,
-        fullWeeks,
-        currentWeekIndex,
-        id,
-    });
-};
-
-const getRoundRow = (
-    originalCells: ReactElement[],
-    round: CalendarRound,
-    index: number,
-    firstMonday: Moment,
-    lastSunday: Moment,
-    currentWeekIndex: number,
-    periodType: PeriodType,
-    campaign: MappedCampaign,
-): ReactElement[] => {
-    // for one round
-    const { id, rounds } = campaign;
-    const previousRound = rounds[index - 1];
-    const startInRange = round.start
-        ? isDateInRange(round.start, firstMonday, lastSunday)
-        : false;
-    const endInRange = round.end
-        ? isDateInRange(round.end, firstMonday, lastSunday)
-        : false;
-
-    let dateUntilNextRound = round.end?.clone().add(round.daysCount, 'days');
-
-    let dateUntilNextRoundInRange =
-        !!dateUntilNextRound &&
-        isDateInRange(dateUntilNextRound, firstMonday, lastSunday);
-
-    const startAndEndInRange = startInRange && endInRange;
-
-    const onlyStartInRange = !endInRange && startInRange;
-
-    const onlyEndInRange = !startInRange && endInRange;
-
-    const isRoundIntervalVisible =
-        !endInRange &&
-        !startInRange &&
-        dateUntilNextRoundInRange &&
-        index === rounds.length - 1;
-
-    const isRoundVisible =
-        startAndEndInRange || onlyStartInRange || onlyEndInRange;
-
-    let cells = addCellsBeforeRound({
-        id,
-        cells: originalCells,
-        round,
-        index,
-        startInRange,
-        previousRound,
-        firstMonday,
-        lastSunday,
-        currentWeekIndex,
-        periodType,
-    });
-    if (round.start && round.end) {
-        if (isRoundVisible) {
-            cells = addRoundCell({
-                campaign,
-                cells,
-                round,
-                firstMonday,
-                lastSunday,
-                startInRange,
-                endInRange,
-                periodType,
-            });
-        }
-
-        if (isRoundVisible || isRoundIntervalVisible) {
-            cells = addBufferCell({
-                id,
-                cells,
-                round,
-                dateUntilNextRoundInRange,
-                lastSunday,
-                endInRange,
-                dateUntilNextRound,
-                firstMonday,
-                rounds,
-                index,
-                periodType,
-            });
-        }
-
-        // if it's last round draw empty cells
-        if (
-            index === rounds.length - 1 &&
-            dateUntilNextRoundInRange &&
-            dateUntilNextRound
-        ) {
-            cells = addRemainingEmptyCells({
-                cells,
-                dateUntilNextRound,
-                lastSunday,
-                campaign,
-                currentWeekIndex,
-            });
-        }
-    } else if (previousRound.end) {
-        // add an additional day for positioning if the next round has a start date
-        if (round.started_at && previousRound.daysCount) {
-            dateUntilNextRound = previousRound.end
-                ?.clone()
-                .add(previousRound.daysCount - 1, 'days');
-            // if next round has empty dates just clone previousRound.end
-        } else {
-            dateUntilNextRound = previousRound.end?.clone();
-        }
-
-        dateUntilNextRoundInRange = isDateInRange(
-            dateUntilNextRound,
-            firstMonday,
-            lastSunday,
-        );
-
-        const previousEndVisible = isDateInRange(
-            previousRound.end,
-            firstMonday,
-            lastSunday,
-        );
-        if (dateUntilNextRoundInRange && !previousEndVisible) {
-            cells = addBufferCell({
-                id,
-                cells,
-                round: { ...round, end: dateUntilNextRound },
-                dateUntilNextRoundInRange,
-                lastSunday,
-                endInRange: false,
-                dateUntilNextRound,
-                firstMonday,
-                rounds,
-                index,
-                periodType,
-            });
-        }
-        cells = addRemainingEmptyCells({
-            cells,
-            dateUntilNextRound,
-            lastSunday,
-            campaign,
-            currentWeekIndex,
-        });
-    }
-    return cells;
-};
-const isStartOfWeek = (date: Moment): boolean => date.isoWeekday() === 1; // 1 is Monday
-const isEndOfWeek = (date: Moment): boolean => date.isoWeekday() === 7; // 7 is Sunday
-
-const getEmptyCellBetweenTwoDates = (
-    startDate: Moment,
-    endDate: Moment,
-    cells: ReactElement[],
-    currentWeekIndex: number,
-): ReactElement[] => {
-    const result: ReactElement[] = [];
-
-    let spans = 0;
-    cells.forEach(c => {
-        spans += c.props.colSpan;
-    });
-
-    spans = Math.floor(spans / 7);
-    // get next monday after startDate
-    let nextMonday;
-    let daysCountBefore = 0;
-    if (!isEndOfWeek(startDate)) {
-        nextMonday = startDate.clone().add(1, 'week').startOf('isoWeek');
-        daysCountBefore = nextMonday.diff(startDate, 'days') || 0;
-    } else {
-        nextMonday = startDate.add(1, 'day').clone();
-    }
-    // get days between startDate and next monday
-    if (daysCountBefore > 0) {
-        spans += 1;
-        result.push(
-            <EmptyCell
-                key={`empty-cell-${startDate}-${daysCountBefore}`}
-                colSpan={daysCountBefore}
-                isCurrentWeek={spans === currentWeekIndex}
-            />,
-        );
-    }
-    // get last sunday before endDate
-    let lastSunday;
-    let daysCountAfter = 0;
-    if (!isStartOfWeek(endDate)) {
-        lastSunday = endDate.clone().subtract(1, 'week').endOf('isoWeek');
-        daysCountAfter = endDate.diff(lastSunday, 'days') || 0;
-    } else {
-        lastSunday = endDate.subtract(1, 'day').clone();
-    }
-    // get weeks between next monday and last sunday
-    const weeksCount =
-        lastSunday.add(1, 'day').diff(nextMonday.subtract(1, 'day'), 'weeks') ||
-        0;
-    if (weeksCount > 0) {
-        Array(weeksCount)
-            .fill(null)
-            .forEach((_, i) => {
-                spans += 1;
-                result.push(
-                    <EmptyCell
-                        key={`empty-cell-week-${weeksCount}-end-${i}`}
-                        colSpan={7}
-                        isCurrentWeek={spans === currentWeekIndex}
-                    />,
-                );
-            });
-    }
-    // get days between last sunday and end date
-    if (daysCountAfter > 0) {
-        spans += 1;
-        result.push(
-            <EmptyCell
-                key={`empty-cell-${endDate}-${daysCountAfter}`}
-                colSpan={daysCountAfter}
-                isCurrentWeek={spans === currentWeekIndex}
-            />,
-        );
-    }
-    return result;
-};
-
-const getSubActivitiesRow = (
-    originalCells: ReactElement[],
-    subActivities: SubActivity[],
-    firstMonday: Moment,
-    lastSunday: Moment,
-    currentWeekIndex: number,
-    campaign: MappedCampaign,
-): ReactElement[] => {
-    // First order subActivities by start date and filter out subActivities that are not in range
-    let cells = [...originalCells];
-    if (subActivities.length > 0) {
-        // Draw cells before first subActivity
-        const firstSubActivity = subActivities[0];
-        const emptyCells = drawEmptyFirstCells({
-            startDate: firstSubActivity.start_date.clone(),
-            firstMonday,
-            currentWeekIndex,
-            id: campaign.id,
-        });
-        cells.push(...emptyCells);
-        // loop into subActivities
-        subActivities.forEach((subActivity, index) => {
-            const startInRange = subActivity.start_date
-                ? isDateInRange(subActivity.start_date, firstMonday, lastSunday)
-                : false;
-            const endInRange = subActivity.end_date
-                ? isDateInRange(subActivity.end_date, firstMonday, lastSunday)
-                : false;
-            // Draw subactivity
-            cells = addSubactivityCell({
-                cells,
-                subActivity,
-                firstMonday,
-                lastSunday,
-                startInRange,
-                endInRange,
-            });
-            // Draw cells after subActivity
-            const nextSubActivity = subActivities[index + 1];
-            if (nextSubActivity) {
-                const nextStartDate = nextSubActivity?.start_date.clone();
-                const emptyBetweenCells = getEmptyCellBetweenTwoDates(
-                    subActivity.end_date,
-                    nextStartDate,
-                    cells,
-                    currentWeekIndex,
-                );
-                cells.push(...emptyBetweenCells);
-            }
-        });
-
-        // Draw cells after last subActivity
-        cells = addRemainingEmptyCells({
-            cells,
-            dateUntilNextRound:
-                subActivities[subActivities.length - 1].end_date,
-            lastSunday,
-            campaign,
-            currentWeekIndex,
-        });
-    }
-
-    return cells;
-};
-
-const getCells = (
-    campaign: MappedCampaign,
-    currentWeekIndex: number,
-    firstMonday: Moment,
-    lastSunday: Moment,
-    periodType: PeriodType,
-): ReactElement[] => {
-    let cells: ReactElement[] = [];
-    const { rounds } = campaign;
-    rounds.forEach((round, index) => {
-        cells = getRoundRow(
-            cells,
-            round,
-            index,
-            firstMonday,
-            lastSunday,
-            currentWeekIndex,
-            periodType,
-            campaign,
-        );
-    });
-    return cells;
-};
-export const getSubActivitiesCells = (
-    campaign: MappedCampaign,
-    currentWeekIndex: number,
-    firstMonday: Moment,
-    lastSunday: Moment,
-): ReactElement[] => {
-    let cells: ReactElement[] = [];
-
-    cells = getSubActivitiesRow(
-        cells,
-        campaign.subActivities,
-        firstMonday,
-        lastSunday,
-        currentWeekIndex,
-        campaign,
-    );
-    return cells;
-};
-export { filterCampaigns, getCalendarData, getCells, mapCampaigns };
